// types/District.ts
export type District =
  | 'Central and Western District'
  | 'Eastern District'
  | 'Southern District'
  | 'Wan Chai District'
  | 'Kowloon City District'
  | 'Kwun Tong District'
  | 'Sham Shui Po District'
  | 'Wong Tai Sin District'
  | 'Yau Tsim Mong District'
  | 'Islands District'
  | 'Kwai Tsing District'
  | 'North District'
  | 'Sai Kung District'
  | 'Sha Tin District'
  | 'Tai Po District'
  | 'Tsuen Wan District'
  | 'Tuen Mun District'
  | 'Yuen Long District';

export const districts: District[] = [
  'Central and Western District',
  'Eastern District',
  'Southern District',
  'Wan Chai District',
  'Kowloon City District',
  'Kwun Tong District',
  'Sham Shui Po District',
  'Wong Tai Sin District',
  'Yau Tsim Mong District',
  'Islands District',
  'Kwai Tsing District',
  'North District',
  'Sai Kung District',
  'Sha Tin District',
  'Tai Po District',
  'Tsuen Wan District',
  'Tuen Mun District',
  'Yuen Long District',
];